+++ src/native.rs
@@ -2,6 +2,10 @@ pub use self::non_utf8::*;
 pub use self::utf8::*;
 
 mod non_utf8 {
+    /// [`Encoding`](crate::Encoding) that is native to the platform during compilation
+    #[cfg(unix)]
+    pub type NativeEncoding = crate::unix::UnixEncoding;
+
     /// [`Path`](crate::Path) that is native to the platform during compilation
     #[cfg(unix)]
     pub type NativePath = crate::unix::UnixPath;
@@ -13,6 +17,10 @@ mod non_utf8 {
     /// [`Component`](crate::Component) that is native to the platform during compilation
     #[cfg(unix)]
     pub type NativeComponent<'a> = crate::unix::UnixComponent<'a>;
+
+    /// [`Encoding`](crate::Encoding) that is native to the platform during compilation
+    #[cfg(windows)]
+    pub type NativeEncoding = crate::windows::WindowsEncoding;
 
     /// [`Path`](crate::Path) that is native to the platform during compilation
     #[cfg(windows)]
@@ -39,6 +47,10 @@ mod non_utf8 {
 }
 
 mod utf8 {
+    /// [`Utf8Path`](crate::Utf8Encoding) that is native to the platform during compilation
+    #[cfg(unix)]
+    pub type Utf8NativeEncoding = crate::unix::Utf8UnixEncoding;
+
     /// [`Utf8Path`](crate::Utf8Path) that is native to the platform during compilation
     #[cfg(unix)]
     pub type Utf8NativePath = crate::unix::Utf8UnixPath;
@@ -50,6 +62,10 @@ mod utf8 {
     /// [`Utf8Component`](crate::Utf8Component) that is native to the platform during compilation
     #[cfg(unix)]
     pub type Utf8NativeComponent<'a> = crate::unix::Utf8UnixComponent<'a>;
+
+    /// [`Utf8Path`](crate::Utf8Encoding) that is native to the platform during compilation
+    #[cfg(windows)]
+    pub type Utf8NativeEncoding = crate::windows::Utf8WindowsEncoding;
 
     /// [`Utf8Path`](crate::Utf8Path) that is native to the platform during compilation
     #[cfg(windows)]