+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-08-23
+
 * Add `normalize` method to `Path` and `Utf8Path` to support resolving `.` and
   `..` across Windows and Unix encodings
 * Add `is_parent` and `is_current` methods to `Component` and `Utf8Component`
   traits with implementations for both Unix and Windows component
   implementations
+* Add `root`, `parent`, and `current` static methods to `Component` and
+  `Utf8Component` traits to support creating the instances from generics
+* Add `absolutize` to both `normalize` the path and - in the case of relative
+  paths - prefix the path with the current working directory
+* Add `to_encoding` to `Path` and `Utf8Path` support converting between the
+  Unix and Windows encoding types
 
 ## [0.3.2] - 2023-03-27
 
@@ -51,7 +59,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release of the library!
 
-[Unreleased]: https://github.com/chipsenkbeil/typed-path/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/chipsenkbeil/typed-path/compare/v0.4.0...HEAD
+[0.3.2]: https://github.com/chipsenkbeil/typed-path/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/chipsenkbeil/typed-path/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/chipsenkbeil/typed-path/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/chipsenkbeil/typed-path/compare/v0.2.1...v0.3.0